DAOS-4049 control: Switch to go-humanize for byte formatting
This library correctly parses both IEC and SI sizes into
uint64 byte counts, and correctly displays byte counts in
the desired formats.

Also fixes some inconsistencies with regard to storage sizes.
Use uint64 byte counts internally and convert for humans at
the edges.

+++ src/control/lib/spdk/nvme.go
@@ -87,7 +87,7 @@ type Controller struct {
 // +inner C.struct_ns_t
 type Namespace struct {
 	ID   int32
-	Size int32
+	Size uint64
 }
 
 // DeviceHealth struct mirrors C.struct_dev_health_t
@@ -266,7 +266,7 @@ func c2GoDeviceHealth(health *C.struct_dev_health_t) *DeviceHealth {
 func c2GoNamespace(ns *C.struct_ns_t) *Namespace {
 	return &Namespace{
 		ID:   int32(ns.id),
-		Size: int32(ns.size),
+		Size: uint64(ns.size),
 	}
 }
